Bound a numeric answer, and stop the aggregate overflowing on old ones
A "number" field declared no range, so 1e308 was a valid number of goals, and
two of them overflowed the running sum inside AggregateScoresForPerson's avg().
That failed the whole query -- every key about the athlete, not just the one
with the big values, since it is one GROUP BY -- and no endpoint deletes a form
instance or an answer, so it stayed failed. The README calls that query the
product's analytical core.

Both halves: validateAnswer bounds every numeric answer at maxAnswerMagnitude,
far outside anything a coach records and far inside float8's ceiling, and the
average accumulates in numeric, because a bound added today cannot reach rows
written yesterday and those rows are unreachable through the API.

See docs/AUDIT-3.md M1.

+++ internal/store/forms.sql.go
@@ -15,7 +15,7 @@ import (
 const aggregateScoresForPerson = `-- name: AggregateScoresForPerson :many
 SELECT ff.key, ff.label,
     count(fa.numeric_value)::bigint      AS samples,
-    avg(fa.numeric_value)::double precision AS average,
+    avg(fa.numeric_value::numeric)::double precision AS average,
     min(fa.numeric_value)::double precision AS minimum,
     max(fa.numeric_value)::double precision AS maximum
 FROM form_instances fi
@@ -45,6 +45,19 @@ type AggregateScoresForPersonRow struct {
 
 // The moat query: cross-instance aggregation of scored fields for one athlete,
 // optionally scoped to a context. Powers readiness means, effort trends, etc.
+//
+// The average accumulates in numeric, not double precision. avg() over float8 sums in
+// float8, so two answers near the type's ceiling overflowed the running sum and the
+// whole query failed — every key at once, not just the one with the big values, because
+// this is one GROUP BY over every answer about the athlete:
+//
+//	ERROR: value out of range: overflow (SQLSTATE 22003)
+//
+// validateAnswer now bounds what can be written (see maxAnswerMagnitude), but rows
+// already in the table are not reachable by any endpoint — nothing in this API deletes a
+// form instance or an answer — so a validation-only fix would leave existing athletes
+// with a permanently 500ing aggregate. numeric has no such ceiling, and the average of
+// values that each fit in float8 always fits in float8 on the way back out.
 func (q *Queries) AggregateScoresForPerson(ctx context.Context, arg AggregateScoresForPersonParams) ([]AggregateScoresForPersonRow, error) {
 	rows, err := q.db.Query(ctx, aggregateScoresForPerson, arg.SubjectPersonID, arg.Context)
 	if err != nil {
